Fix ccapi rpc methods to always pass 8 byte handles instead of sizeof(void*). ccapi server always stores all 8 bytes, whether compiled as 32 bit or 64 bit. If 32 bit, client zero-pads handle when sending and truncates when receiving.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@25265 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/ccapi/server/win/ccs_request_proc.c
@@ -42,8 +42,8 @@ void ccs_rpc_request(
     long*       return_status ) {   /* Return code */
 
     cc_int32        status  = 0;
-    k5_ipc_stream    stream;
-    DWORD*          p       = (DWORD*)(tspHandle);
+    k5_ipc_stream   stream;
+    UINT64*         p       = (UINT64*)(tspHandle);
     WIN_PIPE*       pipe    = NULL;
 #if 0
     cci_debug_printf("%s rpcmsg:%d; UUID:<%s> SST:<%s>", __FUNCTION__, rpcmsg, pszUUID, serverStartTime);
@@ -70,7 +70,7 @@ void ccs_rpc_connect(
     const char* pszUUID,            /* Data buffer */
     long*       return_status ) {   /* Return code */
 
-    DWORD*      p       = (DWORD*)(tspHandle);
+    UINT64*     p       = (UINT64*)(tspHandle);
     WIN_PIPE*   pipe    = ccs_win_pipe_new(pszUUID, *p);
 #if 0
     cci_debug_printf("%s; rpcmsg:%d; UUID: <%s>", __FUNCTION__, rpcmsg, pszUUID);
